+++ apps/gateway/src/users-gateway.controller.ts
@@ -1,15 +1,18 @@
 import { InjectUsersClient } from '@app/common/decorators';
 import { IUsersClient } from '@app/common/users/users.client';
 import {
+  Body,
   Controller,
   Get,
   Headers,
+  Put,
   UnauthorizedException,
 } from '@nestjs/common';
 import type { ClientGrpc } from '@nestjs/microservices';
 import { Metadata } from '@grpc/grpc-js';
 import { GRPC_USERS_SERVICE } from '@app/common/constants';
 import { ApiTags } from '@nestjs/swagger';
+import type { UpdateUserRequest } from '@app/common/users';
 
 @ApiTags('Users')
 @Controller('users')
@@ -25,16 +28,18 @@ export class UsersGatewayController {
 
   @Get('current')
   async getCurrentUser(@Headers('authorization') authHeader: string) {
-    if (!authHeader || !authHeader?.toLowerCase().startsWith('bearer ')) {
-      throw new UnauthorizedException('Token JWT requerido');
-    }
-
-    const jwt = authHeader.substring(7); // Remover 'Bearer '
-
-    // Crear metadata para gRPC
     const metadata = new Metadata();
-    metadata.set('authorization', `Bearer ${jwt}`);
-
+    metadata.set('authorization', authHeader);
     return this.usersService.GetCurrentUser({}, metadata);
+  }
+
+  @Put('current')
+  async updateUser(
+    @Headers('authorization') authHeader: string,
+    @Body() body: UpdateUserRequest,
+  ) {
+    const metadata = new Metadata();
+    metadata.set('authorization', authHeader);
+    return await this.usersService.UpdateUser(body, metadata);
   }
 }