+++ apps/users/src/supabase.adapter.ts
@@ -1,12 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { createClient, SupabaseClient } from '@supabase/supabase-js';
 import { Settings } from './settings';
-import { UserProfile } from '@app/common/users';
+import { UpdateUserRequest, UserProfile } from '@app/common/users';
 
 @Injectable()
 export class SupabaseAdapter {
   private client: SupabaseClient;
-
+  private logger = new Logger(SupabaseAdapter.name);
   constructor() {
     this.client = createClient(
       Settings.supabaseUrl,
@@ -16,6 +16,21 @@ export class SupabaseAdapter {
 
   async getCurrentUser(jwt: string): Promise<UserProfile> {
     const res = await this.client.auth.getUser(jwt);
-    return new UserProfile(res.data.user?.id!, res.data.user?.email!);
+    this.logger.debug(`getCurrentUser: ${JSON.stringify(res)}`);
+    return new UserProfile(
+      res.data.user?.id!,
+      res.data.user?.email!,
+      res.data.user?.user_metadata?.firstName,
+      res.data.user?.user_metadata?.lastName,
+    );
+  }
+
+  async updateUser(id: string, user: UpdateUserRequest): Promise<void> {
+    this.logger.debug(`updateUser: ${JSON.stringify(user)}`);
+    this.client.auth.admin.updateUserById(id, {
+      user_metadata: {
+        ...user,
+      },
+    });
   }
 }