+++ apps/users/src/users.controller.ts
@@ -1,22 +1,22 @@
 import { Controller } from '@nestjs/common';
 import { GrpcMethod, RpcException } from '@nestjs/microservices';
 
-import type { Metadata, ServerUnaryCall } from '@grpc/grpc-js';
+import type { Metadata } from '@grpc/grpc-js';
 import { status } from '@grpc/grpc-js';
 
-import { UserProfile } from '@app/common/users';
-import { UsersService } from './users.service';
+import type { UpdateUserRequest, UserProfile } from '@app/common/users';
 import { GRPC_USERS_SERVICE } from '@app/common/constants';
+import { SupabaseAdapter } from './supabase.adapter';
 
 @Controller()
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(private readonly supabase: SupabaseAdapter) {}
 
   @GrpcMethod(GRPC_USERS_SERVICE, 'GetCurrentUser')
   async getCurrentUser(_: {}, metadata: Metadata): Promise<UserProfile> {
     const authHeader = metadata.get('authorization')[0] as string;
 
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
+    if (!authHeader || !authHeader?.toLowerCase().startsWith('bearer ')) {
       throw new RpcException({
         code: status.UNAUTHENTICATED,
         message: 'Token JWT required',
@@ -25,6 +25,12 @@ export class UsersController {
 
     const jwt = authHeader.substring(7); // Remover 'Bearer '
 
-    return this.usersService.getCurrentUser(jwt);
+    return this.supabase.getCurrentUser(jwt);
+  }
+
+  @GrpcMethod(GRPC_USERS_SERVICE, 'UpdateUser')
+  async updateUser(user: UpdateUserRequest, metadata: Metadata): Promise<void> {
+    const userProfile = await this.getCurrentUser({}, metadata); // Verifica se o usuário está autenticado
+    return await this.supabase.updateUser(userProfile.id, user);
   }
 }