+++ apps/users/src/users.module.ts
@@ -1,11 +1,10 @@
 import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
 import { SupabaseAdapter } from './supabase.adapter';
 
 @Module({
   imports: [],
   controllers: [UsersController],
-  providers: [UsersService, SupabaseAdapter],
+  providers: [SupabaseAdapter],
 })
 export class UsersModule {}