+++ libs/common/src/users/users.client.ts
@@ -1,7 +1,8 @@
 import { Observable } from 'rxjs';
-import { UserProfile } from './users.dto';
+import { UpdateUserRequest, UserProfile } from './users.dto';
 import { Metadata } from '@grpc/grpc-js';
 
 export interface IUsersClient {
   GetCurrentUser(data: {}, metadata: Metadata): Observable<UserProfile>;
+  UpdateUser(data: UpdateUserRequest, metadata: Metadata): Observable<void>;
 }