+++ libs/common/src/users/users.dto.ts
@@ -2,5 +2,9 @@ export class UserProfile {
   constructor(
     public id: string,
     public email: string,
+    public firstName?: string,
+    public lastName?: string,
   ) {}
 }
+
+export type UpdateUserRequest = Omit<UserProfile, 'email' | 'id'>;